Refactoring to organize hard coded data.
