+++ app/src/main/java/com/jordangellatly/hermosahappyhour/model/junkie/JunkieDrinkSpecials.kt
@@ -1,0 +1,26 @@
+package com.jordangellatly.hermosahappyhour.model.junkie
+
+import com.jordangellatly.hermosahappyhour.model.Deal
+
+val junkieFridayHappyHourDrinkSpecials = listOf(
+    Deal(
+        description = "Cadillac Margarita & Not So Old Fashioned",
+        price = "$7",
+        imageUrl = "https://cdn-icons-png.flaticon.com/512/187/187448.png"
+    ),
+    Deal(
+        description = "Cabernet and Chardonnay by the glass",
+        price = "$6",
+        imageUrl = "https://cdn-icons-png.flaticon.com/512/187/187448.png"
+    ),
+    Deal(
+        description = "Draft Bud Light, Buenaveza, Modern Times, & Stash Panda",
+        price = "$4",
+        imageUrl = "https://cdn-icons-png.flaticon.com/512/187/187448.png"
+    ),
+    Deal(
+        description = "Junkie Cup, Mamita-Rita, & Colossal Michelada",
+        price = "$9",
+        imageUrl = "https://cdn-icons-png.flaticon.com/512/187/187448.png"
+    )
+)
