+++ app/src/main/java/com/jordangellatly/hermosahappyhour/model/junkie/JunkieFoodSpecials.kt
@@ -1,0 +1,81 @@
+package com.jordangellatly.hermosahappyhour.model.junkie
+
+import com.jordangellatly.hermosahappyhour.model.Deal
+
+val junkieMondayHappyHourFoodSpecials = listOf(
+    Deal(
+        description = "All Starters",
+        price = "$2 Off",
+        imageUrl = "https://cdn-icons-png.flaticon.com/512/5787/5787016.png"
+    )
+)
+
+val junkieBurgerMondayFoodSpecials = listOf(
+    Deal(
+        description = "Burgers",
+        price = "$5",
+        imageUrl = "https://cdn-icons-png.flaticon.com/512/5787/5787016.png"
+    )
+)
+
+val junkieTuesdayHappyHourFoodSpecials = listOf(
+    Deal(
+        description = "All Starters",
+        price = "$2 Off",
+        imageUrl = "https://cdn-icons-png.flaticon.com/512/5787/5787016.png"
+    ),
+    Deal(
+        description = "AYCE Tacos",
+        price = "$8",
+        imageUrl = "https://cdn-icons-png.flaticon.com/512/5787/5787016.png"
+    )
+)
+
+val junkieWednesdayHappyHourFoodSpecials = listOf(
+    Deal(
+        description = "All Starters",
+        price = "$2 Off",
+        imageUrl = "https://cdn-icons-png.flaticon.com/512/5787/5787016.png"
+    ),
+    Deal(
+        description = "Boneless Wings",
+        price = "$0.50",
+        imageUrl = "https://cdn-icons-png.flaticon.com/512/5787/5787016.png"
+    )
+)
+
+val junkieThursdayHappyHourFoodSpecials = listOf(
+    Deal(
+        description = "All Starters",
+        price = "$2 Off",
+        imageUrl = "https://cdn-icons-png.flaticon.com/512/5787/5787016.png"
+    ),
+    Deal(
+        description = "Personal Size Pizzas",
+        price = "$5",
+        imageUrl = "https://cdn-icons-png.flaticon.com/512/5787/5787016.png"
+    )
+)
+
+val junkieFridayHappyHourFoodSpecials = listOf(
+    Deal(
+        description = "All Starters",
+        price = "$2 Off",
+        imageUrl = "https://cdn-icons-png.flaticon.com/512/5787/5787016.png"
+    ),
+    Deal(
+        description = "Personal Cheese Pizza",
+        price = "$7",
+        imageUrl = "https://cdn-icons-png.flaticon.com/512/5787/5787016.png"
+    ),
+    Deal(
+        description = "Loaded Tots",
+        price = "$8",
+        imageUrl = "https://cdn-icons-png.flaticon.com/512/5787/5787016.png"
+    ),
+    Deal(
+        description = "Carne Asada Fries",
+        price = "$9",
+        imageUrl = "https://cdn-icons-png.flaticon.com/512/5787/5787016.png"
+    )
+)
