+++ app/src/main/java/com/jordangellatly/hermosahappyhour/ui/detail/shared/EventHeader.kt
@@ -9,7 +9,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.jordangellatly.hermosahappyhour.model.thursdayHappyHour
+import com.jordangellatly.hermosahappyhour.model.tower12.tower12ThursdayHappyHour
 import com.jordangellatly.hermosahappyhour.ui.theme.HermosaHappyHourTheme
 
 @Composable
@@ -42,7 +42,7 @@ fun EventHeader(
 @Preview(showBackground = true)
 @Composable
 private fun EventHeaderPreview() {
-    val event = thursdayHappyHour
+    val event = tower12ThursdayHappyHour
     HermosaHappyHourTheme {
         EventHeader(
             weeklyHoursDescription = event.weeklyHoursDescription,
